requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.23
aiomysql==0.2.0
aiosqlite==0.19.0
alembic==1.12.1

# Security
cryptography==41.0.7

# Additional utilities
pydantic==2.5.0
pydantic-settings==2.1.0
